more compiling

+++ storm/plugins/opennlp-me-extractor/src/test/java/io/lumify/opennlpme/OpenNLPMaximumEntropyExtractorGraphPropertyWorkerTest.java
@@ -98,8 +98,8 @@ public class OpenNLPMaximumEntropyExtractorGraphPropertyWorkerTest {
         for (Vertex termMention : termMentions) {
             String title = LumifyProperties.TITLE.getPropertyValue(termMention);
 
-            int start = LumifyProperties.TERM_MENTION_START_OFFSET.getPropertyValue(termMention);
-            int end = LumifyProperties.TERM_MENTION_END_OFFSET.getPropertyValue(termMention);
+            long start = LumifyProperties.TERM_MENTION_START_OFFSET.getPropertyValue(termMention);
+            long end = LumifyProperties.TERM_MENTION_END_OFFSET.getPropertyValue(termMention);
             String conceptType = LumifyProperties.CONCEPT_TYPE.getPropertyValue(termMention);
 
             if (title.equals("Bob Robértson")) {
